+++ decorators/index.js
@@ -1,0 +1,1 @@
+export { default as ctrlWrapper } from "./ctrlWrapper.js";
