+++ package.json
@@ -12,8 +12,10 @@
   "dependencies": {
     "cors": "2.8.5",
     "cross-env": "7.0.3",
+    "dotenv": "^16.3.1",
     "express": "4.17.1",
     "joi": "^17.9.2",
+    "mongoose": "^7.4.0",
     "morgan": "1.10.0",
     "nanoid": "^4.0.2"
   },
